Changed i64 to i32

+++ src/main.rs
@@ -38,8 +38,8 @@ struct Opt {
     opening: Option<u32>,
 }
 
-fn break_duration_to_min(d: Duration) -> (i64, f64) {
-    let minutes = (d.as_secs_f64() / 60.0) as i64;
+fn break_duration_to_min(d: Duration) -> (i32, f64) {
+    let minutes = (d.as_secs_f64() / 60.0) as i32;
     let seconds = d.as_secs_f64() - minutes as f64 * 60.0;
     (minutes, seconds)
 }
